refactor: 파일 위치 조정 (icons, views)
