+++ public/icons/note-view.png
