+++ src/views/todo/todo-item/TodoItem.tsx
@@ -1,8 +1,8 @@
 import { TodoResponse } from "@/types/todo";
 
-import { ActionIcon } from "../../atoms/action-icon/ActionIcon";
-import { TodoCheckbox } from "../../atoms/todo-checkbox/TodoCheckbox";
-import { TodoTitle } from "../../atoms/todo-title/TodoTitle";
+import { TodoCheckbox } from "../todo-checkbox/TodoCheckbox";
+import { ActionIcon } from "./action-icon/ActionIcon";
+import { TodoTitle } from "./todo-title/TodoTitle";
 
 interface TodoItemProps {
   todo: TodoResponse["todos"][number];
@@ -20,7 +20,7 @@ export function TodoItem({
   return (
     <li className="group mb-2 flex h-6 w-full last:mb-0">
       <TodoCheckbox
-        done={todo.done}
+        checked={todo.done}
         onToggle={() => handleToggleTodo(todo.id, todo.done)}
       />
       <div className="flex w-full min-w-0 items-center justify-between">
